change to see if it works

+++ server/src/seeds/cleanDb.ts
@@ -3,9 +3,14 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    const model = models[modelName];
+    if (!model || !model.db) {
+      throw new Error(`Model or database connection for ${modelName} is undefined.`);
+    }
+
+    const modelExists = await model.db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
